rename rep_labels folder
